feat: Core of the Projects Page "Treasure Room"
Directory for the treasure-room page.
Overall theme set and rough timeline piece started.
Need to apply theme to timeline feature and fill with correct information.
Allow the each project to cover the screen on click.
Possibly create a list of projects at the top so users don't need to scroll down every time.